pyproject.toml
[project]
name = "stuffr"
version = "0.1.0"
description = "Platform for posting announcements"
authors = [
    {name = "Vladislav",email = "barabayne@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "alembic (>=1.14.1,<2.0.0)",
    "uuid7 (>=0.1.0,<0.2.0)",
    "fastapi (>=0.115.8,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pytest (>=8.3.4,<9.0.0)",
    "ruff (>=0.9.5,<0.10.0)",
    "pre-commit (>=4.1.0,<5.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "bcrypt (>=4.2.1,<5.0.0)",
    "pydantic[email] (>=2.10.6,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
